Preserve the frozen frontier source behind the publication extension
