+++ simllm/deploy/__init__.py
@@ -61,23 +61,25 @@ from simllm.deploy.frontier import (
     CandidateFrontier,
     ExternalAnchor,
     FrontierPoint,
-    FrontierPublicationComparison,
     FrontierRecord,
-    FrontierSelection,
-    FrontierSelectionSegment,
     PointClass,
-    PublicationBasis,
-    PublishedThreshold,
     ScanInputs,
-    ThresholdInterval,
-    compare_published_frontier,
-    frontier_at_threshold,
     frontier_record_from_json,
     frontier_record_to_json,
     pareto_front,
     prepare_plot_v3,
     scan,
     weak_dominance_pareto,
+)
+from simllm.deploy.frontier_publication import (
+    FrontierPublicationComparison,
+    FrontierSelection,
+    FrontierSelectionSegment,
+    PublicationBasis,
+    PublishedThreshold,
+    ThresholdInterval,
+    compare_published_frontier,
+    frontier_at_threshold,
 )
 from simllm.deploy.ladder import (
     FRONTIER_LADDER_RECORD_SCHEMA,
